fix: resolve all clippy warnings and lint issues
- Remove unused imports and fix redundant patterns
- Fix unnecessary casts and needless borrows
- Add allow(dead_code) for Stripe webhook structs
- Move regex compilation outside loop in validate-upload
- Fix redundant .ok() patterns across all lambda functions

+++ lambdas/sketch-builder/src/main.rs
@@ -221,7 +221,7 @@ async fn build_sketch_tracks(
         .send()
         .await?;
 
-    let original_size = obj.content_length().unwrap_or(0) as i64;
+    let original_size = obj.content_length().unwrap_or(0);
     let data = obj.body.collect().await?.into_bytes();
 
     // Save to temporary file
